minoda commit .classpath
